+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - ReLinker code cleanup
 - Gradle wrapper version updated to 2.10
 - Fixed ANR on disposing `GifTextureView` - [#240](https://github.com/koral--/android-gif-drawable/issues/240)
+- Added `GifDecoder` for access GIF frames without `Drawable` or `View` - [#206](https://github.com/koral--/android-gif-drawable/issues/206)
 
 ### 1.1.11
 - `MultiCallback` now accepts `Drawable.Callback`s, not only `View`s